refactor(deepseek-v41): own DSpark draft construction
