+++ nemo_automodel/components/models/deepseek_v41/config.py
@@ -16,10 +16,29 @@
 
 from __future__ import annotations
 
+import copy
 import math
-from typing import Any
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any, Protocol
 
 from transformers import AutoTokenizer, PretrainedConfig, PreTrainedTokenizerFast
+
+if TYPE_CHECKING:
+    from nemo_automodel.components.models.deepseek_v41.dspark import DeepseekV41DSparkModel
+
+
+class _DSparkDraftOptions(Protocol):
+    """Declarative DSpark settings supplied by the generic training recipe."""
+
+    num_draft_layers: int
+    target_layer_ids: list[int]
+    block_size: int
+    num_anchors: int
+    mask_token_id: int
+    markov_rank: int
+    markov_head_type: str
+    confidence_head_alpha: float
+    confidence_head_with_markov: bool
 
 
 class DeepseekV41TextConfig(PretrainedConfig):
@@ -430,3 +449,105 @@ class DeepseekV41Config(PretrainedConfig):
         if not isinstance(tokenizer, PreTrainedTokenizerFast):
             raise TypeError("DeepSeek-V4.1 Engram requires the checkpoint's original fast tokenizer")
         return tokenizer
+
+    def build_dspark_draft(self, options: _DSparkDraftOptions) -> DeepseekV41DSparkModel:
+        """Build the model-owned DSpark training adapter.
+
+        Args:
+            options: Declarative draft settings supplied by the DSpark recipe.
+
+        Returns:
+            Native DeepSeek V4.1 DSpark model ready for device placement.
+        """
+        return DeepseekV41DSparkConfig.from_target(self.text_config, options).build()
+
+
+@dataclass(frozen=True)
+class DeepseekV41DSparkConfig:
+    """Validated declarative configuration for the released V4.1 DSpark module."""
+
+    text_config: DeepseekV41TextConfig
+    num_draft_layers: int
+    target_layer_ids: tuple[int, ...]
+    block_size: int
+    num_anchors: int
+    mask_token_id: int
+    markov_rank: int
+    markov_head_type: str
+    confidence_head_alpha: float
+    confidence_head_with_markov: bool
+
+    @classmethod
+    def from_target(
+        cls,
+        text_config: DeepseekV41TextConfig,
+        options: _DSparkDraftOptions,
+    ) -> DeepseekV41DSparkConfig:
+        """Create an immutable DSpark config from target and recipe settings.
+
+        Args:
+            text_config: Released DeepSeek V4.1 text checkpoint configuration.
+            options: Declarative DSpark recipe settings.
+
+        Returns:
+            Immutable model-owned DSpark configuration.
+        """
+        return cls(
+            text_config=text_config,
+            num_draft_layers=int(options.num_draft_layers),
+            target_layer_ids=tuple(int(layer_id) for layer_id in options.target_layer_ids),
+            block_size=int(options.block_size),
+            num_anchors=int(options.num_anchors),
+            mask_token_id=int(options.mask_token_id),
+            markov_rank=int(options.markov_rank),
+            markov_head_type=str(options.markov_head_type),
+            confidence_head_alpha=float(options.confidence_head_alpha),
+            confidence_head_with_markov=bool(options.confidence_head_with_markov),
+        )
+
+    def build(self) -> DeepseekV41DSparkModel:
+        """Validate the released contract and build its training adapter.
+
+        Returns:
+            Newly initialized DeepSeek V4.1 DSpark training model.
+
+        Raises:
+            ValueError: If recipe settings disagree with the released checkpoint.
+        """
+        from nemo_automodel.components.models.deepseek_v41.dspark import DeepseekV41DSparkModel
+
+        expected = {
+            "num_draft_layers": self.text_config.num_nextn_predict_layers,
+            "block_size": self.text_config.dspark_block_size,
+            "mask_token_id": self.text_config.dspark_noise_token_id,
+            "markov_rank": self.text_config.dspark_markov_rank,
+        }
+        for name, expected_value in expected.items():
+            actual_value = getattr(self, name)
+            if actual_value != expected_value:
+                raise ValueError(
+                    f"DeepSeek V4.1 DSpark requires {name}={expected_value} from the target checkpoint, "
+                    f"got {actual_value}"
+                )
+        if list(self.target_layer_ids) != self.text_config.dspark_target_layer_ids:
+            raise ValueError(
+                "DeepSeek V4.1 DSpark target_layer_ids must match the target checkpoint: "
+                f"expected {self.text_config.dspark_target_layer_ids}, got {list(self.target_layer_ids)}"
+            )
+        if self.num_anchors <= 0:
+            raise ValueError(f"DeepSeek V4.1 DSpark requires num_anchors > 0, got {self.num_anchors}")
+        if not math.isfinite(self.confidence_head_alpha) or self.confidence_head_alpha < 0:
+            raise ValueError(
+                "DeepSeek V4.1 DSpark requires confidence_head_alpha to be finite and non-negative, "
+                f"got {self.confidence_head_alpha}"
+            )
+        if self.markov_head_type != "vanilla":
+            raise ValueError(f"DeepSeek V4.1 DSpark requires markov_head_type='vanilla', got {self.markov_head_type!r}")
+        if not self.confidence_head_with_markov:
+            raise ValueError("DeepSeek V4.1 DSpark requires confidence_head_with_markov=true")
+
+        draft_config = copy.deepcopy(self.text_config)
+        draft_config.architectures = ["DeepseekV41DSparkModel"]
+        draft_config.dspark_num_anchors = self.num_anchors
+        draft_config.dspark_enable_confidence_head = self.confidence_head_alpha > 0
+        return DeepseekV41DSparkModel(draft_config)
