+++ nemo_automodel/recipes/llm/train_dspark.py
@@ -79,7 +79,6 @@ from nemo_automodel.components.optim.optimizer import build_optimizer
 from nemo_automodel.components.speculative.dspark.common import validate_target_layer_ids
 from nemo_automodel.components.speculative.dspark.config import (
     build_deepseek_v4_draft_config,
-    build_deepseek_v41_draft_config,
     build_gemma4_draft_config,
     build_glm_5_2_draft_config,
     build_kimi_k3_draft_config,
@@ -87,6 +86,7 @@ from nemo_automodel.components.speculative.dspark.config import (
 )
 from nemo_automodel.components.speculative.dspark.core import DSparkStepMetrics, DSparkTrainerModule
 from nemo_automodel.components.speculative.dspark.registry import (
+    ModelOwnedDSparkProvider,
     build_target_layer_ids,
     resolve_dspark_draft_spec,
 )
@@ -1020,13 +1020,22 @@ class TrainDSparkRecipe(BaseRecipe):
         confidence_head_alpha = float(recipe_cfg.get("confidence_head_alpha", 1.0))
         markov_rank = int(recipe_cfg.get("markov_rank", 256))
 
-        if (
-            is_deepseek_v4_target
-            or is_deepseek_v41_target
-            or is_glm_5_2_target
-            or is_gemma4_target
-            or is_minimax_m3_target
-            or is_kimi_k3_target
+        draft_model = None
+        if isinstance(target_config, ModelOwnedDSparkProvider):
+            margs = _DraftArgs(
+                num_draft_layers=draft_num_hidden_layers,
+                target_layer_ids=target_layer_ids,
+                block_size=self.block_size,
+                num_anchors=self.num_anchors,
+                mask_token_id=self.mask_token_id,
+                markov_rank=markov_rank,
+                markov_head_type=str(recipe_cfg.get("markov_head_type", "vanilla")),
+                confidence_head_alpha=confidence_head_alpha,
+                confidence_head_with_markov=bool(recipe_cfg.get("confidence_head_with_markov", True)),
+            )
+            draft_model = target_config.build_dspark_draft(margs)
+        elif (
+            is_deepseek_v4_target or is_glm_5_2_target or is_gemma4_target or is_minimax_m3_target or is_kimi_k3_target
         ):
             # Gemma4, DeepSeek V4, GLM-5.2, MiniMax M3, and Kimi K3 drafts share one typed
             # draft-config builder that takes the same DSpark model-args bundle.
@@ -1045,8 +1054,6 @@ class TrainDSparkRecipe(BaseRecipe):
                 # The V4 draft is always dense and fixes _attn_implementation to "sdpa"
                 # inside the builder, so it is not overridden by attention_backend.
                 draft_config_obj = build_deepseek_v4_draft_config(target_config, margs)
-            elif is_deepseek_v41_target:
-                draft_config_obj = build_deepseek_v41_draft_config(target_config, margs)
             elif is_glm_5_2_target:
                 # The GLM draft is always dense and fixes _attn_implementation to "sdpa"
                 # inside the builder, so it is not overridden by attention_backend.
@@ -1085,8 +1092,10 @@ class TrainDSparkRecipe(BaseRecipe):
             draft_config_obj = Qwen3Config.from_dict(draft_config)
             draft_config_obj._attn_implementation = attention_backend
 
-        draft_cls = resolve_dspark_draft_spec(architectures).draft_cls
-        self.draft_model = draft_cls(draft_config_obj).to(device=self.device)
+        if draft_model is None:
+            draft_cls = resolve_dspark_draft_spec(architectures).draft_cls
+            draft_model = draft_cls(draft_config_obj)
+        self.draft_model = draft_model.to(device=self.device)
         cast_model_to_dtype(self.draft_model, self.compute_dtype)
         if self.packed_sequence_size > 0 and type(self.draft_model).__name__ != "Qwen3DSparkModel":
             # Only the Qwen3 draft forward threads the packing metadata so far; the
